feat: add output print feature
- UserInputClass is now integration outputPrint feature
- Modify printMessage (final static)
